ml/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "garden-ml"
version = "0.1.0"
description = "Smart Tomato Garden ML pipeline (features+XGB and baseline DL) with reproducibility, metrics, and inference API."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "joblib>=1.3",
  "opencv-python-headless>=4.8",
  "Pillow>=10.0",
  "mahotas>=1.4",
  "xgboost>=2.0",
  "tqdm>=4.66",
  "loguru>=0.7.2",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "albumentations>=1.4.0",
  "scikit-image>=0.22",
  "flask>=3.0",
  "gunicorn>=21.2",
  "prometheus-client>=0.20",
  "requests>=2.32",
  "mlflow>=2.12",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6", "mypy>=1.8"]
dl = [
  "torch>=2.2",
  "torchvision>=0.17",
  "timm>=1.0.0",
]

[project.scripts]
garden-ml-augment = "garden_ml.training.augment:main"
garden-ml-train = "garden_ml.training.train:main"
garden-ml-eval = "garden_ml.training.evaluate:main"
garden-ml-deep = "garden_ml.training.deep_baseline:main"
garden-ml-serve = "garden_ml.inference.api_flask:main"

[tool.hatch.build.targets.wheel]
packages = ["src/garden_ml"]

[tool.ruff]
line-length = 120
